refactor(dashboards): extract ServiceCard from ServicesDashboard
Phase 1 of #1072's scoping plan. The 240-LOC inline ServiceCard
component moves into packages/frontend/src/components/ServiceCard.tsx
as a pure presentational component. The dashboard passes in the
action callbacks (onMonitor, onEdit, onActions, onEditLink, onDelete,
onRestart, onContainerLogs/Terminal/Actions) plus the live-state
derived bits (httpsDomains, attachNodeContext) it owns; the card
stays oblivious to multi-node bookkeeping.

Behaviour delta: zero. The JSX, useMemo for dedupedPorts, useCallback
for ensureContainerContext, all the data-testid hooks, and the failed-
state nudge all transfer verbatim. The wider prop interface is a
deliberate refactor cost - ServiceCard had 11 closures over the
dashboard's scope; making them explicit props is the price of the
move.

File line count: ServicesDashboard.tsx 1944 → 1717 (-227). New file
ServiceCard.tsx is 304 LOC. Lint warning count unchanged at 428;
ServicesDashboard's max-lines warning ratchets from 1785 → 1717 and
the inner-component count drops accordingly.

Phase 1 of #1072. Issue stays open for Phase 2 (YamlEditorModal)
and Phase 3 (StackDeployModal).
